+++ includes/admin/class-wc-payplus-admin-settings.php
@@ -357,6 +357,15 @@ class WC_PayPlus_Admin_Settings
                     'class' => 'payplus-documents'
                 );
                 $settings[$section][] = array(
+                    'name' => __("Brand UID - Development/Sandbox", 'payplus-payment-gateway'),
+                    'id' => 'payplus_invoice_option[payplus_invoice_brand_uid_sandbox]',
+                    'type' => 'text',
+                    'class' => 'payplus_invoice_brand_uid',
+                    'desc' => __('Set development - brand UID from which the system will issue the documents (Leave blank if you only have one brand)', 'payplus-payment-gateway'),
+                    'desc_tip' => true,
+                    'class' => 'payplus-documents'
+                );
+                $settings[$section][] = array(
                     'name' => __("Website code", 'payplus-payment-gateway'),
                     'id' => 'payplus_invoice_option[payplus_website_code]',
                     'type' => 'text',