+++ includes/admin/class-wc-payplus-admin.php
@@ -25,6 +25,7 @@ class WC_PayPlus_Admin_Payments extends WC_PayPlus_Gateway
     public $allSettings;
     public $transactionType;
     private $_wpnonce;
+    public $api_test_mode;
 
     /**
      * @return null
@@ -69,6 +70,7 @@ class WC_PayPlus_Admin_Payments extends WC_PayPlus_Gateway
         }
 
         $this->transactionType = $this->get_option('transaction_type');
+        $this->api_test_mode = $this->get_option('api_test_mode');
 
         $this->payPlusInvoice = new PayplusInvoice();
         $payPlusInvoiceOptions = get_option('payplus_invoice_option');
@@ -1832,7 +1834,7 @@ class WC_PayPlus_Admin_Payments extends WC_PayPlus_Gateway
         $transactionType = $this->get_option('transaction_type');
 
         wp_enqueue_style('payplus', PAYPLUS_PLUGIN_URL . 'assets/css/admin.min.css', [], PAYPLUS_VERSION);
-        wp_register_script('payplus-admin-payment', PAYPLUS_PLUGIN_URL . '/assets/js/admin-payments.min.js', ['jquery'], time(), true);
+        wp_register_script('payplus-admin-payment', PAYPLUS_PLUGIN_URL . '/assets/js/admin-payments.js', ['jquery'], time(), true);
         wp_localize_script(
             'payplus-admin-payment',
             'payplus_script_admin',
@@ -1848,6 +1850,7 @@ class WC_PayPlus_Admin_Payments extends WC_PayPlus_Gateway
                 "payplus_enabled_payment" => $enabled,
                 "payplusTransactionType" => $this->transactionType,
                 "payplus_invoice" => $isInvoice,
+                "testMode" => $this->api_test_mode,
                 "payplus_refund_error" => __('Incorrect amount or amount greater than amount that can be refunded', 'payplus-payment-gateway'),
                 "menu_option" => WC_PayPlus::payplus_get_admin_menu(wp_create_nonce('menu_option')),
                 "payplus_refund_club_amount" => wp_create_nonce('payplus_refund_club_amount'),