+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * Fix    - Express checkout activation.
 * Add    - Multipass icons will change with fade in and out on the checkout page.
 * Add    - Multipass select clubs settings in plugin settings.
+* Add    - Brand UID for Sandbox/Development mode.
 
 
 ##  [7.0.8] - 2024-07-15 - (Shinobi)